requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
openai>=1.0.0
aiohttp>=3.8.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # For PostgreSQL support